+++ packages/rangelink-vscode-extension/src/navigation/index.ts
@@ -1,3 +1,4 @@
+export * from './pickFilenameCandidate';
 export * from './RangeLinkDocumentProvider';
 export * from './RangeLinkNavigationHandler';
 export * from './RangeLinkTerminalProvider';